Code cleaned up and separated into components
