+++ src/new.jsx
@@ -1,0 +1,18 @@
+import React from 'react'
+import CodeEditor from './components/code-editor/code-editor'
+
+
+const New = () => {
+
+  // const { code, setCode } = React.useContext(EditorContext);
+
+  return (
+    <>
+      <CodeEditor />
+
+    </>
+
+  )
+}
+
+export default New
